refactor(server): collapse persistence into core SqliteStore (#117)
Move conversations / extraction_jobs / events tables into the core
SqliteWorker pipeline so the whole codebase has one persistence model.

Before:
- apps/server/src/persistence.rs (742) opens its own Connection per call,
  hand-rolls 3-table SQL, lives outside the worker singleton
- apps/server/src/application/ports.rs defines repo traits with
  `Result<T, String>` for server-only consumption
- conversations records/enums live in apps/server/src/models.rs

After:
- packages/core/src/conversation/ owns record types + repository traits
  (async, InfraResult<T>); usable from cli/desktop/mirror too
- packages/core/src/infra/sqlite/conversation_ops.rs runs the SQL inside
  the existing worker thread; PRAGMA / WAL / busy_timeout / FK
  automatically apply (also resolves #98 CR-4 by construction)
- SqliteStore now impls ConversationRepository + JobRepository +
  EventRepository; AppState wires a single Arc<SqliteStore> into all four
  slots (store / doc_store / conversation_repo / job_repo / event_repo)
- All callers in apps/server/src/application/* + extraction.rs + handlers.rs
  now `.await` the repo calls and map InfraError -> String for their
  per-handler error enums

Tests:
- 7 ServerPersistence tests rewritten against SqliteStore in
  conversation_ops.rs, exercising the trait surface
- cargo test --workspace: 252 passed, 0 failed

Net delta: server -770 lines, core +800 lines (incl. tests). No
backward-compatibility shim; ServerPersistence is fully removed per
the project's no-compat policy.

Refs #98 (CR-4 PRAGMA bypass — closed by construction)
Closes #117
