+++ apps/server/src/application/job.rs
@@ -38,7 +38,8 @@ pub async fn get_extraction_job(
     let job = state
         .job_repo
         .find_job_by_id(&job_id)
-        .map_err(ExtractionJobError::Internal)?
+        .await
+        .map_err(|err| ExtractionJobError::Internal(err.to_string()))?
         .ok_or_else(|| ExtractionJobError::NotFound("Job not found".to_string()))?;
     Ok(ExtractionJobResult { job })
 }
